moved PoolActionsView inside of Wallet routes
