--- src/renderer/routes/history/history.test.ts
@@ -1,12 +1,0 @@
-import { base } from './history'
-
-describe('History routes', () => {
-  describe('base routes', () => {
-    it('template', () => {
-      expect(base.template).toEqual('/history')
-    })
-    it('path', () => {
-      expect(base.path()).toEqual('/history')
-    })
-  })
-})
