--- src/renderer/routes/history/history.ts
@@ -1,8 +1,0 @@
-import { Route } from '../types'
-
-export const base: Route<void> = {
-  template: '/history',
-  path() {
-    return this.template
-  }
-}
