--- src/renderer/routes/history/index.ts
@@ -1,1 +1,0 @@
-export * from './history'
