+++ src/renderer/views/wallet/PoolActionsHistory/PoolActionsHistoryView.styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { BackLink as BackLinkUI } from '../../components/uielements/backLink'
+import { BackLink as BackLinkUI } from '../../../components/uielements/backLink'
 
 export const BackLink = styled(BackLinkUI)`
   width: max-content;
