+++ src/renderer/views/wallet/PoolActionsHistory/index.ts
@@ -1,0 +1,1 @@
+export * from './PoolActionsHistoryView'
